move styling from workspace to panel components; add navigation to canvas panel via props
