+++ client/src/components/ActivityPanels/helpers.js
