+++ client/src/views/StudentLogin/StudentLogin.js
@@ -93,7 +93,7 @@ export default function StudentLogin(props) {
                     <button id='link-button' onClick={removeStudent}>Remove a student</button>
                 </div>
                 {/*error && <div style={{ color: 'red' }}>{error}</div>*/}
-                <button type='submit' onClick={() => handleLogin(studentIds)}>Enter</button>
+                <button id='login-button' type='submit' onClick={() => handleLogin(studentIds)}>Enter</button>
             </div>
         </div>
     )
