+++ client/src/views/StudentLogin/StudentLogin.less
@@ -121,7 +121,7 @@
   font-size: 1.4em;
 }
 
-button {
+#login-button {
   background: darken(#colors[tertiary], 30%);
   width: 20%;
   display: block;
